+++ NEWS.md
@@ -1,7 +1,8 @@
-# governor 0.1.0.9002  2024-07-09
+# governor 0.1.0.9003  2024-07-09
 
 * Windows now uses `nanosleep()`
 * `timer` has now moved to C
+* Option to enable/disable tivers and governors
 
 # governor 0.1.0  2024-06-29